+++ src/cell/mod.rs
@@ -1,5 +1,7 @@
 //! Shareable mutable containers.
 
+mod cell;
 mod unsafe_cell;
 
+pub use self::cell::Cell;
 pub use self::unsafe_cell::{ConstPtr, MutPtr, UnsafeCell};