+++ src/rt/atomic.rs
@@ -1,34 +1,60 @@
 //! An atomic cell
 //!
-//! # Coherence
+//! See the CDSChecker paper for detailed explanation.
 //!
-//! The coherence rules are implemented as follows.
+//! # Modification order implications (figure 7)
 //!
-//! - Read-Read:
+//! - Read-Read Coherence:
 //!
 //!   On `load`, all stores are iterated, finding stores that were read by
 //!   actions in the current thread's causality. These loads happen-before the
 //!   current load. The `modification_order` of these happen-before loads are
 //!   joined into the current load's `modification_order`.
 //!
-//! - Write-Read:
+//! - Write-Read Coherence:
 //!
 //!   On `load`, all stores are iterated, finding stores that happens-before the
 //!   current thread's causality. The `modification_order` of these stores are
 //!   joined into the current load's `modification_order`.
 //!
-//! - Read-Write:
+//! - Read-Write Coherence:
 //!
 //!   On `store`, find all existing stores that were read in the current
 //!   thread's causality. Join these stores' `modification_order` into the new
 //!   store's modification order.
 //!
-//! - Write-Write:
+//! - Write-Write Coherence:
 //!
 //!   The `modification_order` is initialized to the thread's causality. Any
 //!   store that happened in the thread causality will be earlier in the
 //!   modification order.
+//!
+//! - Seq-cst/MO Consistency:
+//!
+//! - Seq-cst Write-Read Coherence:
+//!
+//! - RMW/MO Consistency: Subsumed by Write-Write Coherence?
+//!
+//! - RMW Atomicity:
+//!
+//!
+//! # Fence modification order implications (figure 9)
+//!
+//! - SC Fences Restrict RF:
+//! - SC Fences Restrict RF (Collapsed Store):
+//! - SC Fences Restrict RF (Collapsed Load):
+//! - SC Fences Impose MO:
+//! - SC Fences Impose MO (Collapsed 1st Store):
+//! - SC Fences Impose MO (Collapsed 2st Store):
+//!
+//!
+//! # Fence Synchronization implications (figure 10)
+//!
+//! - Fence Synchronization
+//! - Fence Synchronization (Collapsed Store)
+//! - Fence Synchronization (Collapsed Load)
 
+use crate::rt::execution::Execution;
 use crate::rt::location::{self, Location, LocationSet};
 use crate::rt::object;
 use crate::rt::{
@@ -39,6 +65,8 @@ use std::cmp;
 use std::marker::PhantomData;
 use std::sync::atomic::Ordering;
 use std::u16;
+
+use tracing::trace;
 
 #[derive(Debug)]
 pub(crate) struct Atomic<T> {
@@ -137,27 +165,47 @@ struct FirstSeen([u16; MAX_THREADS]);
 
 /// Implements atomic fence behavior
 pub(crate) fn fence(ordering: Ordering) {
-    use std::sync::atomic::Ordering::Acquire;
-
-    assert_eq!(
-        ordering, Acquire,
-        "only Acquire fences are currently supported"
-    );
-
-    rt::synchronize(|execution| {
-        // Find all stores for all atomic objects and, if they have been read by
-        // the current thread, establish an acquire synchronization.
-        for state in execution.objects.iter_mut::<State>() {
-            // Iterate all the stores
-            for store in state.stores_mut() {
-                if !store.first_seen.is_seen_by_current(&execution.threads) {
-                    continue;
-                }
-
-                store.sync.sync_load(&mut execution.threads, ordering);
-            }
-        }
+    rt::synchronize(|execution| match ordering {
+        Ordering::Acquire => fence_acq(execution),
+        Ordering::Release => fence_rel(execution),
+        Ordering::AcqRel => fence_acqrel(execution),
+        Ordering::SeqCst => fence_seqcst(execution),
+        Ordering::Relaxed => panic!("there is no such thing as a relaxed fence"),
+        order => unimplemented!("unimplemented ordering {:?}", order),
     });
+}
+
+fn fence_acq(execution: &mut Execution) {
+    // Find all stores for all atomic objects and, if they have been read by
+    // the current thread, establish an acquire synchronization.
+    for state in execution.objects.iter_mut::<State>() {
+        // Iterate all the stores
+        for store in state.stores_mut() {
+            if !store.first_seen.is_seen_by_current(&execution.threads) {
+                continue;
+            }
+
+            store
+                .sync
+                .sync_load(&mut execution.threads, Ordering::Acquire);
+        }
+    }
+}
+
+fn fence_rel(execution: &mut Execution) {
+    // take snapshot of cur view and record as rel view
+    let active = execution.threads.active_mut();
+    active.released = active.causality;
+}
+
+fn fence_acqrel(execution: &mut Execution) {
+    fence_acq(execution);
+    fence_rel(execution);
+}
+
+fn fence_seqcst(execution: &mut Execution) {
+    fence_acqrel(execution);
+    execution.threads.seq_cst_fence();
 }
 
 impl<T: Numeric> Atomic<T> {
@@ -166,6 +214,8 @@ impl<T: Numeric> Atomic<T> {
         rt::execution(|execution| {
             let state = State::new(&mut execution.threads, value.into_u64(), location);
             let state = execution.objects.insert(state);
+
+            trace!(?state, "Atomic::new");
 
             Atomic {
                 state,
@@ -193,6 +243,8 @@ impl<T: Numeric> Atomic<T> {
             // Get the store to return from this load.
             let index = execution.path.branch_load();
 
+            trace!(state = ?self.state, ?ordering, "Atomic::load");
+
             T::from_u64(state.load(&mut execution.threads, index, location, ordering))
         })
     }
@@ -208,6 +260,8 @@ impl<T: Numeric> Atomic<T> {
 
             // An unsync load counts as a "read" access
             state.track_unsync_load(&execution.threads);
+
+            trace!(state = ?self.state, "Atomic::unsync_load");
 
             // Return the value
             let index = index(state.cnt - 1);
@@ -227,6 +281,8 @@ impl<T: Numeric> Atomic<T> {
             // An atomic store counts as a read access to the underlying memory
             // cell.
             state.track_store(&execution.threads);
+
+            trace!(state = ?self.state, ?ordering, "Atomic::store");
 
             // Do the store
             state.store(
@@ -264,6 +320,8 @@ impl<T: Numeric> Atomic<T> {
             // Get the store to use for the read portion of the rmw operation.
             let index = execution.path.branch_load();
 
+            trace!(state = ?self.state, ?success, ?failure, "Atomic::rmw");
+
             state
                 .rmw(
                     &mut execution.threads,
@@ -290,6 +348,8 @@ impl<T: Numeric> Atomic<T> {
             // Verify the mutation may happen
             state.track_unsync_mut(&execution.threads);
             state.is_mutating = true;
+
+            trace!(state = ?self.state, "Atomic::with_mut");
 
             // Return the value of the most recent store
             let index = index(state.cnt - 1);
@@ -688,6 +748,7 @@ impl State {
                 let mo_i = store_i.modification_order;
                 let mo_j = store_j.modification_order;
 
+                // TODO: this sometimes fails
                 assert_ne!(mo_i, mo_j);
 
                 if mo_i < mo_j {