+++ src/sync/arc.rs
@@ -1,6 +1,7 @@
 use crate::rt;
 
-use std::ops;
+use std::pin::Pin;
+use std::{mem, ops};
 
 /// Mock implementation of `std::sync::Arc`.
 #[derive(Debug)]
@@ -9,6 +10,7 @@ pub struct Arc<T> {
 }
 
 #[derive(Debug)]
+#[repr(C)]
 struct Inner<T> {
     // This must be the first field to make into_raw / from_raw work
     value: T,
@@ -28,10 +30,30 @@ impl<T> Arc<T> {
         Arc { inner }
     }
 
+    /// Constructs a new `Pin<Arc<T>>`.
+    pub fn pin(data: T) -> Pin<Arc<T>> {
+        unsafe { Pin::new_unchecked(Arc::new(data)) }
+    }
+
     /// Gets the number of strong (`Arc`) pointers to this value.
     pub fn strong_count(_this: &Self) -> usize {
         unimplemented!("no tests checking this? DELETED!")
         // this.inner.ref_cnt.load(SeqCst)
+    }
+
+    /// Increments the strong reference count on the `Arc<T>` associated with the
+    /// provided pointer by one.
+    pub unsafe fn increment_strong_count(ptr: *const T) {
+        // Retain Arc, but don't touch refcount by wrapping in ManuallyDrop
+        let arc = mem::ManuallyDrop::new(Arc::<T>::from_raw(ptr));
+        // Now increase refcount, but don't drop new refcount either
+        let _arc_clone: mem::ManuallyDrop<_> = arc.clone();
+    }
+
+    /// Decrements the strong reference count on the `Arc<T>` associated with the
+    /// provided pointer by one.
+    pub unsafe fn decrement_strong_count(ptr: *const T) {
+        mem::drop(Arc::from_raw(ptr));
     }
 
     /// Returns a mutable reference to the inner value, if there are
@@ -53,11 +75,14 @@ impl<T> Arc<T> {
 
     /// Consumes the `Arc`, returning the wrapped pointer.
     pub fn into_raw(this: Self) -> *const T {
-        use std::mem;
-
-        let ptr = &*this as *const _;
+        let ptr = Self::as_ptr(&this);
         mem::forget(this);
-        ptr as *const T
+        ptr
+    }
+
+    /// Provides a raw pointer to the data.
+    pub fn as_ptr(this: &Self) -> *const T {
+        std::sync::Arc::as_ptr(&this.inner) as *const T
     }
 
     /// Constructs an `Arc` from a raw pointer.